Add 1/4-block support for truecolor mode
This was fun, since in 1/4-block mode you have 4 addressable sub-blocks,
but only 2 colors to use on them.

Solution is to try each block pattern, and blend together the colors of
the pixels in the foreground part to make the foreground color, and the
colors of the pixels in the background part to make the background
color. Each pattern is then compared, and the pattern with the least
overall "error" is chosen.